Restored old volume tracking module
Removed some unnecessary variables.
The general principal of the protocol works now, the pipette does go down for every pipetting step, the problem is that it goes down too fast so that it thinks the mix is finished long before it actually is finished (it goes to deep into the liquid).