Updated test_volume_tracking.py
test_volume_tracking_internal_module_old.py update: changed back to the old working protocol. We now have the version that has working volume tracking but goes down a bit too far in the conical part of the tube. We decided that for now this is fine, and implemented that at the moment that the pipette thinks it is at the bottom it starts pipetting from the bottom so that it does go on untill the plate is filled.
We also implemented that the tip is changed every 2 rows and that the blow out (should be) at the same location as the aspiration.

Co-Authored-By: Maartje <78266071+maartje-git@users.noreply.github.com>