Updated calibration check & start illu primer 12S
Updated calibration check so that we could calibrate tWALL-E with the 20 ul tips. All labware and pipet tips that we needed to aliquot the illumina primers + the 20 ul pipet tips were included so that we didn't have to calibrate any labware in the Illumina protocol.

Started with protocol for illu primers 12S but didn't really added anything yet except a description.