Made a protocol for 12S illumina PCR
Made a protocol for the illumina PCR of 12S, with some improvements compared to the illumina PCR protocol for 6 strips (which I will implement there as well),

Problem here was that I wanted to pipette into 5 full strips + 2 tubes, so I couldn't use the way I used for the 6 strips, because the two separate wells were a problem. So now I created 2 lists, 1 with columns and 1 with wells, and then used the double flattened loop to append the wells from the columns and a normal for loop for the wells to append those wells to ONE list. Then in the for loop for the pipetting I used that list that I created.

Next problem with this was that I couldn't get a new tip for every column anymore...

This is solved by enumerating the list and using an if elif statement to pick up a new tip for the first well, and than after every 8 wells dropping the tip and than picking up a new tip.